+++ .storybook/preview.js
@@ -1,9 +1,9 @@
 import * as nextImage from "next/image";
 import "../icomoon/style.css";
 import "../styles/globals.css";
-import "../styles/fonts.css";
 import "../styles/forms.css";
 import "../styles/menu.css";
+import "./fonts.css";
 
 import i18n from "./i18n.js";
 import { I18nProviderWrapper } from "./i18nextProviderWrapper";