+++ pages/signup.js
@@ -68,6 +68,26 @@ export default function Signup(props) {
         });
         return errors;
       }),
+    confirmEmail: Joi.string()
+      .email({ tlds: { allow: false } })
+      .required()
+      .equal(Joi.ref("email"))
+      .error((errors) => {
+        errors.forEach((error) => {
+          switch (error.code) {
+            case "any.required":
+              error.message = t("emailRequired");
+              break;
+            case "string.email":
+              error.message = t("errorEmail");
+            case "any.only":
+              error.message = t("emailError");
+            default:
+              break;
+          }
+        });
+        return errors;
+      }),
     yearOfBirthRange: Joi.string()
       .invalid(yearOptions[0].id)
       .required()
@@ -168,6 +188,9 @@ export default function Signup(props) {
   const [email, setEmail] = useState("");
   const [emailError, setEmailError] = useState("");
 
+  const [confirmEmail, setConfirmEmail] = useState("");
+  const [confirmEmailError, setConfirmEmailError] = useState("");
+
   const [yearOfBirthRange, setYearOfBirthRange] = useState("");
   const [yearOfBirthRangeError, setYearOfBirthRangeError] = useState("");
 
@@ -191,6 +214,7 @@ export default function Signup(props) {
   const [minorityGroupOther, setMinorityGroupOther] = useState("");
 
   const [incomeLevel, setIncomeLevel] = useState("");
+  const [publicServant, setPublicServant] = useState("");
 
   const [agreeToConditions, setAgreeToConditions] = useState("");
   const [agreeToConditionsError, setAgreeToConditionsError] = useState("");
@@ -219,6 +243,7 @@ export default function Signup(props) {
   const handlerClearData = (e) => {
     e.preventDefault();
     setEmailError("");
+    setConfirmEmailError("");
     setLanguageError("");
     setYearOfBirthRangeError("");
     setProvinceError("");
@@ -226,6 +251,7 @@ export default function Signup(props) {
     setAgreeToConditionsError("");
 
     setEmail("");
+    setConfirmEmail("");
     setYearOfBirthRange("");
     setLanguage("");
     setGender("");
@@ -237,6 +263,7 @@ export default function Signup(props) {
     setMinorityGroup([]);
     setMinorityGroupOther("");
     setIncomeLevel("");
+    setPublicServant("");
     setAgreeToConditions("");
   };
 
@@ -244,6 +271,7 @@ export default function Signup(props) {
     e.preventDefault();
     // clear out error values
     await setEmailError("");
+    await setConfirmEmailError("");
     await setLanguageError("");
     await setYearOfBirthRangeError("");
     await setProvinceError("");
@@ -255,6 +283,7 @@ export default function Signup(props) {
     // compile data into one object
     const formData = {
       email,
+      confirmEmail,
       yearOfBirthRange,
       language,
       province,
@@ -267,6 +296,7 @@ export default function Signup(props) {
       minorityGroup,
       minorityGroupOther,
       incomeLevel,
+      publicServant,
       agreeToConditions,
     };
 
@@ -291,6 +321,7 @@ export default function Signup(props) {
       // map error message setters to field names so that they can be called dynamically
       const errorSetFunctions = {
         email: setEmailError,
+        confirmEmail: setConfirmEmailError,
         language: setLanguageError,
         yearOfBirthRange: setYearOfBirthRangeError,
         province: setProvinceError,
@@ -536,6 +567,18 @@ export default function Signup(props) {
                 onChange={setEmail}
                 boldLabel={true}
                 describedby="emailDoNoInclude"
+                required
+              />
+              <TextField
+                className="mb-10"
+                label={t("confirmEmail")}
+                type="email"
+                name="confirmEmail"
+                id="confirmEmail"
+                error={confirmEmailError}
+                value={confirmEmail}
+                onChange={setConfirmEmail}
+                boldLabel={true}
                 required
               />
               <SelectField
@@ -1036,6 +1079,31 @@ export default function Signup(props) {
                   checked={incomeLevel === "preferNotToAnswer"}
                   onChange={(checked, name, value) => setIncomeLevel(value)}
                   value="preferNotToAnswer"
+                />
+              </fieldset>
+
+              <fieldset className="mb-16">
+                <legend className="block leading-tight text-sm lg:text-p font-body mb-5 font-bold">
+                  {t("formPublicServant")}{" "}
+                  <span className="inline text-form-input-gray text-sm lg:text-p not-italic">
+                    {t("optional")}
+                  </span>
+                </legend>
+                <RadioField
+                  label={t("yes")}
+                  id="publicServantYes"
+                  name="publicServant"
+                  checked={publicServant === "yes"}
+                  onChange={(checked, name, value) => setPublicServant(value)}
+                  value="yes"
+                />
+                <RadioField
+                  label={t("no")}
+                  id="publicServantNo"
+                  name="publicServant"
+                  checked={publicServant === "no"}
+                  onChange={(checked, name, value) => setPublicServant(value)}
+                  value="no"
                 />
               </fieldset>